+++ src/math/centerOfMass.h
@@ -5,7 +5,7 @@
 
 static inline float polygonArea(const std::vector<glm::vec2>& path)
 {
-    int size = path.size();
+    int size = static_cast<int>(path.size());
     if (size < 3) return 0;
 
     float a = 0;
