+++ src/math/triangulate.h
@@ -12,7 +12,7 @@ public:
 
     static bool process(const Path& input, Indices& output)
     {
-        int n = input.size();
+        int n = static_cast<int>(input.size());
         if (n < 3)
             return false;
 
@@ -74,7 +74,7 @@ private:
     static float area(const Path &input)
     {
         float result = 0;
-        int p0 = input.size() - 1;
+        int p0 = static_cast<int>(input.size()) - 1;
         for(unsigned int p1=0; p1<input.size(); p1++)
         {
             result += input[p0].x * input[p1].y - input[p1].x * input[p0].y;
